requirements.txt
discord.py
python-dotenv
simpy
requests
matplotlib